cleaning up add, create, delete scipts

+++ add_s3_trigger.py
@@ -1,4 +1,10 @@
-#!/bin/env python
+#!/usr/bin/env python
+"""
+This script adds a trigger to the zappa_settings file
+The Trigger is for the UPLOAD_BUCKET defined in the bucket_config file
+And calls our app/events.py s3_uploadTrigger() function
+"""
+
 import json, datetime, os, bucketConfig
 from pprint import pprint
 from shutil import copyfile
@@ -7,16 +13,26 @@ from shutil import copyfile
 zappa_settings = 'zappa_settings.json'
 # Name of the function that we will call with trigger
 triggerFunction = "app.events.s3_uploadTrigger"
+# Name of the event on s3 bucket
+s3event = 's3:ObjectCreated:Post'
+# set the Upload bucket's ARN
+bucketArn = "arn:aws:s3:::%s" % bucketConfig.UPLOAD_BUCKET
 
 # Create backup dir
-backupDir = 'backup'
+backupDir = './backups'
 if not os.path.exists(backupDir):
     os.mkdir(backupDir)
 
+# Make sure that the zappa_settings file exists
+if not os.path.isfile(zappa_settings):
+    print "Required file: %s doesnt exist. quitting" % zappa_settings
+    print "Try running 'zappa init' or get a copy from: %s" %backupDir
+    exit(1)
+
 # Backup the zappa settings file
 datestamp = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-backupFile = "%s/%s.%s.json" %(backupDir, zappa_settings[:-5], datestamp)
-print "Backing up zappa_settings to %s " %backupFile
+backupFile = "%s/%s.%s" %(backupDir, zappa_settings, datestamp)
+print "Backing up zappa_settings to %s" %backupFile
 copyfile( zappa_settings, backupFile )
 
 # Read the zappa setting json file
@@ -27,15 +43,20 @@ with open(zappa_settings) as data_file:
 data["dev"]["events"] = [{
             "function": "%s" % triggerFunction,
             "event_source": {
-                  "arn":  "arn:aws:s3:::%s" %bucketConfig.UPLOAD_BUCKET,
+                  "arn":  "%s" %bucketArn,
                   "events": [
-                    "s3:ObjectCreated:Post"
+                    "%s" % s3event
                   ]
                }
             }]
 
-# Write th json to the zappa config file
-print "Writing up zappa_settings to %s " % zappa_settings
+# Tell the user what we are doing
+print "Configurting s3 event: '%s'" %s3event
+print " On s3 bucket: '%s'" %bucketArn
+print " That triggers function: '%s'" %triggerFunction
+
+# Write the json to the zappa config file
+print "Writing these settings to %s" % zappa_settings
 data_file = open(zappa_settings,'w')
 data_file.write(json.dumps(data, indent=4))
 data_file.close()
